requirements.txt
google-adk
kaggle
pandas